move remaining things
