+++ bookshelf/packtest/__init__.py
@@ -1,1 +1,6 @@
 """Python wrapper to run PackTest for a given version."""
+
+from .assets import Assets
+from .runner import Runner
+
+__all__ = ["Assets", "Runner"]
