+++ bookshelf/plugins/bundle_packs.py
@@ -4,14 +4,14 @@ from functools import cache
 
 from beet import Context, subproject
 
-from core.definitions import MODULES, MODULES_DIR
+from bookshelf.definitions import MODULES, MODULES_DIR
 
 __path__ = ""
 
 def beet_default(ctx: Context) -> None:
     """Include all modules."""
     for mod in MODULES or []:
-        config = {"directory": f"../{mod}", "extend": "module.json"}
+        config = {"directory": f"{MODULES_DIR}/{mod}", "extend": "module.json"}
         ctx.require(subproject(config))
 
 @cache
@@ -21,6 +21,6 @@ def __getattr__(tag: str) -> Callable[[Context], None]:
         for mod in MODULES or []:
             meta = json.loads((MODULES_DIR / mod / "module.json").read_text("utf-8"))
             if tag in meta.get("meta", {}).get("tags", []):
-                config = {"directory": f"../{mod}", "extend": "module.json"}
+                config = {"directory": f"{MODULES_DIR}/{mod}", "extend": "module.json"}
                 ctx.require(subproject(config))
     return plugin
