+++ build.py
@@ -1,7 +1,8 @@
 import shutil
 
 
-def pdm_build_initialize(context):
+def pdm_build_initialize(context) -> None: # noqa: ANN001
+    """Add modules to the bookshelf directory when building project."""
     context.ensure_build_dir()
     shutil.copytree(
         "./modules",
