+++ modules/bs.block/templates/items_table.jinja
@@ -1,3 +1,3 @@
-{% include 'core/header.jinja' %}
+{% include 'bookshelf/header.jinja' %}
 # This file was automatically generated, do not edit it
 data modify storage bs:const block.items set value {{ items }}
