+++ pyproject.toml
@@ -34,7 +34,7 @@ docs = [
   "sphinx-design>=0.6.1",
   "sphinx-minecraft>=1.0.2",
   "sphinx-togglebutton>=0.3.2",
-  "sphinx>=8.2.0",
+  "sphinx>=8.2.1",
 ]
 
 [project.scripts]
